+++ plainweb/src/admin/layout.tsx
@@ -1,4 +1,4 @@
-import type Html from "@kitajs/html";
+import type { Html } from "@kitajs/html";
 
 export default function RootLayout(
   props: Html.PropsWithChildren<{