+++ plainweb/tsup.config.ts
@@ -2,7 +2,7 @@ import { defineConfig } from "tsup";
 
 export default defineConfig({
   target: "node20",
-  format: ["cjs", "esm"],
+  format: ["cjs"],
   entry: ["src/index.ts"],
   dts: true,
   outDir: "dist",