+++ template/CHANGELOG.md
@@ -1,5 +1,12 @@
 # my-app
 
+## 1.0.9
+
+### Patch Changes
+
+- Updated dependencies
+  - plainweb@0.0.9
+
 ## 1.0.8
 
 ### Patch Changes